requirements.txt
pkg_resources
fastapi
uvicorn[standard]
websockets
httpx
sounddevice
soundfile
python-dotenv
python-multipart
aiofiles
requests
webrtcvad
